Split out scope code from html module
